change of class name and tests refactor
